+++ pillow_heif/private.py
@@ -38,22 +38,32 @@ def get_pure_stride(mode: str, width: int):
 MODE_CONVERT = {
     # source_mode: {target_mode: convert_function,}
     "BGRA;16": {"RGBA;10": lib.convert_bgra16_to_rgba10, "RGBA;12": lib.convert_bgra16_to_rgba12},
+    "BGRa;16": {"RGBa;10": lib.convert_bgra16_to_rgba10, "RGBa;12": lib.convert_bgra16_to_rgba12},
     "BGR;16": {"RGB;10": lib.convert_bgr16_to_rgb10, "RGB;12": lib.convert_bgr16_to_rgb12},
     "RGBA;16": {"RGBA;10": lib.convert_rgba16_to_rgba10, "RGBA;12": lib.convert_rgba16_to_rgba12},
+    "RGBa;16": {"RGBa;10": lib.convert_rgba16_to_rgba10, "RGBa;12": lib.convert_rgba16_to_rgba12},
     "RGB;16": {"RGB;10": lib.convert_rgb16_to_rgb10, "RGB;12": lib.convert_rgb16_to_rgb12},
     "L;16": {"L;10": lib.convert_i16_to_i10, "L;12": lib.convert_i16_to_i12},
     "I;16": {"L;10": lib.convert_i16_to_i10, "L;12": lib.convert_i16_to_i12},
     "I;16L": {"L;10": lib.convert_i16_to_i10, "L;12": lib.convert_i16_to_i12},
     "RGBA;12": {"RGBA;16": lib.convert_rgba12_to_rgba16, "BGRA;16": lib.convert_rgba12_to_bgra16},
+    "RGBa;12": {"RGBa;16": lib.convert_rgba12_to_rgba16, "BGRa;16": lib.convert_rgba12_to_bgra16},
     "RGB;12": {"RGB;16": lib.convert_rgb12_to_rgb16, "BGR;16": lib.convert_rgb12_to_bgr16},
     "RGBA;10": {"RGBA;16": lib.convert_rgba10_to_rgba16, "BGRA;16": lib.convert_rgba10_to_bgra16},
+    "RGBa;10": {"RGBa;16": lib.convert_rgba10_to_rgba16, "BGRa;16": lib.convert_rgba10_to_bgra16},
     "RGB;10": {"RGB;16": lib.convert_rgb10_to_rgb16, "BGR;16": lib.convert_rgb10_to_bgr16},
     "BGRA": {"RGBA": lib.convert_bgra_rgba},
+    "BGRa": {"RGBa": lib.convert_bgra_rgba},
     "BGR": {"RGB": lib.convert_bgr_rgb},
     "RGBA": {
         "BGRA": lib.convert_bgra_rgba,
         "RGBA;16": lib.convert_rgba_to_rgba16,
         "BGRA;16": lib.convert_rgba_to_bgra16,
+    },
+    "RGBa": {
+        "BGRa": lib.convert_bgra_rgba,
+        "RGBa;16": lib.convert_rgba_to_rgba16,
+        "BGRa;16": lib.convert_rgba_to_bgra16,
     },
     "RGB": {"BGR": lib.convert_bgr_rgb, "RGB;16": lib.convert_rgb_to_rgb16, "BGR;16": lib.convert_rgb_to_bgr16},
 }