+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Support for images with a "premultiplied" alpha channel (Pillow does not fully support these).
+
 ### Changed
 
 ### Fixed