+++ CHANGELOG.md
@@ -6,9 +6,11 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Support for images with `premultiplied` alpha channel.
+- Support for images with a "premultiplied" alpha channel (Pillow does not fully support these).
 
 ### Changed
+
+- `libaom` updated from `3.4.0` to `3.5.0` version. [changelog](https://aomedia.googlesource.com/aom/+/refs/tags/v3.5.0)
 
 ### Fixed